+++ src/connections/dto/Connection.dto.ts
